Split lib.rs
This is mostly organization.
